dev-requirements.txt
cryptography~=39.0
grpcio==1.41.0
grpcio-tools==1.41.0
minknow-api-service-plugin
pre-commit
pytest
